pyproject.toml
[tool.poetry]
name = "slack-time-localization-bot"
version = "0.1.0-dev0"
description = "Detect temporal expressions in Slack messages (_tomorrow at 5 pm_) and translate them for readers in other timezones."
authors = ["Philipp Hossner <philipp.hossner@posteo.de>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
slack-bolt = "^1.18.1"
cachetools = "^5.3.2"
pyduckling-native-phihos = "^0.2.0"
lingua-language-detector = "^2.0.2"
pendulum = "^3.0.0"
regex = "^2023.12.25"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = { extras = ["d"], version = "^23.11.0" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
